p2/backend/requirements.txt
# ─── Core utilities (install first) ──────────────────────────────────────────
setuptools>=65.0.0
wheel>=0.40.0

# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.115.14
uvicorn[standard]==0.34.3
python-multipart==0.0.6
websockets==12.0

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy==2.0.48

# ─── Scientific Computing ─────────────────────────────────────────────────────
# numpy 2.x required for Python 3.13+
numpy>=2.0.0
scipy>=1.13.0

# ─── Audio Processing ─────────────────────────────────────────────────────────
# NOTE: FFmpeg must also be installed on your system (see README)
openai-whisper==20250625
librosa==0.11.0
soundfile==0.13.1

# ─── Document Processing ──────────────────────────────────────────────────────
PyPDF2==3.0.1
python-docx==1.1.0
# python-magic is optional; file validation falls back gracefully without it
# Requires libmagic system library — skip unless you need strict MIME checking
# python-magic==0.4.27

# ─── Environment ──────────────────────────────────────────────────────────────
python-dotenv==1.0.1
email-validator==2.3.0

# ─── Authentication & Security ────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
# passlib 1.7.4 + bcrypt>=4.0 causes AttributeError on Python 3.12.
# Fix: pin bcrypt to 4.0.1 AND install passlib with the patch below.
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ─── Google / OAuth ───────────────────────────────────────────────────────────
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# ─── Gemini AI ────────────────────────────────────────────────────────────────
google-generativeai==0.8.6

# ─── MFA / QR ─────────────────────────────────────────────────────────────────
pyotp==2.9.0
qrcode[pil]==7.4.2

# ─── Video Processing ─────────────────────────────────────────────────────────
# opencv is already installed; mediapipe + deepface enable advanced facial analysis
# Install: pip install mediapipe deepface tensorflow-cpu pandas retina-face --no-deps
# opencv-python==4.9.0.80  # already installed
# mediapipe>=0.10.0         # already installed
# deepface==0.0.99          # install with --no-deps
# tensorflow-cpu            # required by deepface
# pandas                    # required by deepface
# retina-face==0.0.17       # required by deepface (install --no-deps)

# ─── FFmpeg (bundled — used for audio conversion when system FFmpeg is absent) ─
imageio-ffmpeg==0.6.0
